% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaUnderSensFdr.R
\name{AreaUnderSensFdr}
\alias{AreaUnderSensFdr}
\title{Calculate FDR difference and AUC.}
\usage{
AreaUnderSensFdr(sensfdr, cutoff = 0.25)
}
\arguments{
\item{sensfdr}{the output object or SensFdr() which contains the information
of sensitivity, real FDR, and reported FDR.}

\item{cutoff}{the reported FDR cutoff for calculating the Fdr.Diff and AUC.
Ranges of reported FDR greater than this cutoff will not be considered.}
}
\value{
a data frame. Columns are c('Fdr.Diff','Area').
}
\description{
This function is used to calculate the FDR difference (Fdr.Diff) and
Area Under Sensitifity-Real_FDR Curve (AUC).
}
\examples{
a = AreaUnderSensFdr(sensfdr = data.frame(Sensitivity = abs(rnorm(10)), Real_FDR = abs(rnorm(10)), Reported_FDR = abs(rnorm(10))))
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
