% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchPropTest.R
\name{branchPropTest}
\alias{branchPropTest}
\title{Perform branch proportion test.}
\usage{
branchPropTest(data, design)
}
\arguments{
\item{data}{a dataframe or matrix where the each row is a branch, each column is a sample (patient). Values are the branch proportion. This data can be the second element of the output (a list) of evaluate_uncertainty().}

\item{design:}{a data frame. Number of rows should be the same as the number of unique samples. Each row is a sample. First column is the sample names. Second column is the covariate realization valuels for each of the samples that needs to be tested on.}
}
\value{
a vector of p.values.
}
\description{
This function is used to perform branch proportion test.
}
\examples{
a <- branchPropTest(data = matrix(rnorm(24), nrow =3), design = data.frame(sample = paste0('BM', seq(1,8)), sex = c(rep(1,4), rep(0,4))))
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
