% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_prop_test.R
\name{cell_prop_test}
\alias{cell_prop_test}
\title{Perform cell proportion test.}
\usage{
cell_prop_test(
  cellanno,
  pseudotime,
  design = NULL,
  permuiter = 100,
  EMmaxiter = 100,
  ncores = detectCores()
)
}
\arguments{
\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{permuiter}{the number of permutation. Value will be passed to the "permuiter" of lamian.test().}

\item{EMmaxiter}{the number of EM iterations. Value will be passed to the "EMmaxiter" of lamian.test().}

\item{ncores}{number of cores. if ncores > 1, then implement in parallel mode.}

\item{design:}{a matrix. Number of rows should be the same as the number of unique samples. Rownames are sample names. First column is the intercept (all 1), second column is the covariate realization valuels for each of the samples.}
}
\value{
a list. A lamian.test output object except there is one more element of statistics in the list.
}
\description{
This function is used to perform cell proportion test.
}
\examples{
data(mandata)
a <- cellPropTest(cellanno = mandata$cellanno, pseudotime = mandata$pseudotime, design = mandata$design, ncores = 1, permuiter = 2, EMmaxiter = 3)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
