% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellPropTest.R
\name{cellPropTest}
\alias{cellPropTest}
\title{Perform cell proportion test.}
\usage{
cellPropTest(
  cellanno,
  pseudotime,
  design = NULL,
  ncores = detectCores(),
  test.type = "variable",
  testvar = 2
)
}
\arguments{
\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{ncores}{number of cores. if ncores > 1, then implement in parallel mode.}

\item{test.type}{the type of test. test.type = 'Time', use TCD test. test.type = 'Variable', use XCD test.}

\item{testvar}{a numeric number indicating the column in the design matrix that needs to be tested while controlling for other columns (not intercept). Default is 2. testvar = 2 means the second column in the design needs to be tested.}

\item{design:}{a matrix. Number of rows should be the same as the number of unique samples. Rownames are sample names. First column is the intercept (all 1), second column is the covariate realization valuels for each of the samples.}
}
\value{
a list. A lamian_test output object except there is one more element of statistics in the list.
}
\description{
This function is used to perform cell proportion test.
}
\examples{
data(mandata)
a <- cellPropTest(cellanno = mandata$cellanno, pseudotime = mandata$pseudotime, design = mandata$design, ncores = 1)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
