% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGene.R, R/cluster_gene.R
\name{cluster_gene}
\alias{cluster_gene}
\title{Cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).}
\usage{
cluster_gene(
  testobj,
  gene,
  k,
  k.auto = FALSE,
  type = "time",
  method = "kmeans",
  scale.difference = FALSE
)

cluster_gene(
  testobj,
  gene,
  k,
  k.auto = FALSE,
  type = "time",
  method = "kmeans",
  scale.difference = FALSE
)
}
\arguments{
\item{testobj}{object returned from lamian.test().}

\item{gene}{a character vector of gene names. It can be of length 1 or > 1.}

\item{k}{a numeric number. The number of clusters. Only useful when k.auto = FALSE.}

\item{k.auto}{logical. If FALSE (default), users need to specify k as the number of clusters. If TRUE, k will be automatically determined by elbow's method.}

\item{type}{One of c('Time', 'Variable').}

\item{method}{the clustering method, either "kmeans" for k-means clustering or "hierarchical" for hierarchical clustering.}

\item{scale.difference}{logical. If FALSE, then do not standarize the group difference, but only scale by the maximum value of the group difference absolute values. If TRUE, then standardize the group difference before doing the clustering.}
}
\value{
a plot

a plot
}
\description{
This function is used to cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).

This function is used to support the external function clusterGene(). cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).
}
\examples{
data(mantestobj)
clu = clusterGene(testobj = mantestobj, k = 4)
}
\author{
Wenpin Hou <whou10@jhu.edu>

Wenpin Hou <whou10@jhu.edu>
}
