% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGene.R
\name{clusterGene}
\alias{clusterGene}
\title{Cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).}
\usage{
clusterGene(
  testobj,
  gene,
  type = "variable",
  k.auto = FALSE,
  k = 5,
  method = "kmeans",
  scale.difference = F,
  seed = 12345
)
}
\arguments{
\item{testobj}{object returned from lamian_test().}

\item{gene}{a character vector of gene names. It can be of length 1 or > 1.}

\item{type}{A character denoting the population fit is on "Time" or "Variable".  Default is "Time". Case insensitive.}

\item{k.auto}{logical. If FALSE (default), users need to specify k as the number of clusters. If TRUE, k will be automatically determined by elbow's method.}

\item{k}{a numeric number. The number of clusters. Only useful when k.auto = FALSE.}

\item{method}{the clustering method. Options include 'kmeans'(default), 'hierarchical', 'louvain', and 'GMM'.  "kmeans" (default) for k-means clustering,  "hierarchical" for hierarchical clustering, "louvain" for Louvain clustering, and "GMM" for Model-based clustering. Note: "GMM" is very slow.}

\item{scale.difference}{logical. If FALSE, then do not standarize the group difference, but only scale by the maximum value of the group difference absolute values. If TRUE, then standardize the group difference before doing the clustering.}
}
\value{
a plot
}
\description{
This function is used to cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).
}
\examples{
data(mantestobj)
clu = clusterGene(testobj = mantestobj, k = 4)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
