% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbranch.R
\name{findbranch}
\alias{findbranch}
\title{Enumerate pseudotime tree branches}
\usage{
findbranch(mst, order, origin)
}
\arguments{
\item{mst}{minimum spanning tree object. It is MSTtree element in the output (a list) of TSCAN::exprmclust().}

\item{order}{a list of cell orders. Each element is a vector of cells. It is the output (a list) of TSCAN::TSCANorder().}

\item{origin}{a numeric number indicating the origin cluster.}
}
\value{
a list of branch
}
\description{
This function will take the  minimum spanning tree object from the output (a list) of TSCAN::exprmclust() and enumerate all tree branches.
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
