% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfunc_h5.R
\name{fitfunc_h5}
\alias{fitfunc_h5}
\title{Perform model fitting in one permutation setting.}
\usage{
fitfunc_h5(
  iter,
  diffType = "overall",
  gene = NULL,
  testvar = testvar,
  test.type = "Time",
  expr = expr,
  cellanno = cellanno,
  pseudotime = pseudotime,
  design = design,
  EMmaxiter = 100,
  EMitercutoff = 0.05,
  verbose.output = F,
  ncores = 1
)
}
\arguments{
\item{iter}{a numeric number indicating which iteration it is in a permutation test. If iter = 1, it is  fitting the original data (not permuted).}

\item{diffType}{one of overall, meanDiff, or trendDiff.}

\item{gene}{a vector of gene names. It is used when users want to externally specify the genes (or specify the order or genes).}

\item{testvar}{a numeric number indicating the column in the design matrix that needs to be tested while controlling for other columns (not intercept). Its value when calls the function will be used. e.g. testvar = 2 means the second column in the design needs to be tested.}

\item{test.type}{the type of test. One of c('Time', 'Variable). Case insensitive.}

\item{expr}{hdf5 file path (including file name). This hdf5 file is the gene expression.}

\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{EMmaxiter}{an integer indicating the number of iterations in the EM algorithm.}

\item{EMitercutoff}{a numeric number indicating the log-likelihood cutoff applied to stop the EM algorithm}

\item{verbose.output}{logical. If TRUE, print intermediate information.}

\item{ncores}{the number of cores to be used. If ncores > 1, it will be implemented in parallel mode.}

\item{design:}{a matrix. Number of rows should be the same as the number of unique samples. Rownames are sample names. First column is the intercept (all 1), second column is the covariate realization valuels for each of the samples.}
}
\value{
a list of two elements. The first is for full model, while the second is for null model. Each element is a list containing data and parameter estimates.
}
\description{
This function is used to perform model fitting in one permutation setting.
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
