% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpt.R
\name{fitpt}
\alias{fitpt}
\title{Perform the model fitting.}
\usage{
fitpt(
  expr,
  cellanno,
  pseudotime,
  design,
  testvar = testvar,
  maxknotallowed = 10,
  EMmaxiter = 100,
  EMitercutoff = 0.05,
  ncores = 1,
  model = 3,
  knotnum = NULL
)
}
\arguments{
\item{expr}{gene by cell expression matrix. The expression values should have been library-size-normalized and log-transformed. They can either be imputed or non-imputed.}

\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{maxknotallowed}{a numeric number. Max number of knots applied in the B-spline fitting.}

\item{EMmaxiter}{an integer indicating the number of iterations in the EM algorithm.}

\item{EMitercutoff}{a numeric number indicating the log-likelihood cutoff applied to stop the EM algorithm}

\item{ncores}{number of cores used when performing the spline fitting for genes.}

\item{model}{a numeric number (1,2,or 3) indicating which model from the nesting models will be used for fitting.  Model 0 is implemented by the function fitpt.m0().}

\item{knotnum}{If NULL (default), this function will automatically select the optimal number of knots for each gene. If specified by a numeric vector whose names are gene names, this function will used the specified number of knots for those genes. This argument is used to speed up the fitting if the number of knots has been known.}

\item{design:}{a matrix. Number of rows should be the same as the number of unique samples. Rownames are sample names. First column is the intercept (all 1), second column is the covariate realization valuels for each of the samples.}
}
\value{
a list containing parameters and number of knots that obtained from the model for genes.
}
\description{
This function is used to perform the model fitting.
}
\examples{
data(mandata)
a = fitpt(expr = mandata$expr, cellanno = mandata$cellanno, pseudotime = mandata$pseudotime, design = mandata$design, testvar = 1,  maxknotallowed=5, EMmaxiter=10, EMitercutoff=10, ncores=1, model = 1)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
