% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpt_m0_h5.R
\name{fitpt_m0_h5}
\alias{fitpt_m0_h5}
\title{Perform the model 0 (beta0 for intercept base only) fitting.}
\usage{
fitpt_m0_h5(
  expr,
  pseudotime,
  design,
  targetgene = NULL,
  boot = NULL,
  EMmaxiter = 100,
  EMitercutoff = 0.05
)
}
\arguments{
\item{expr}{hdf5 path dir (including file name).}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{targetgene}{the gene names need to be fitted. if NULL, then read in h5 file "feature" of the first sample.}

\item{EMmaxiter}{an integer indicating the number of iterations in the EM algorithm.}

\item{EMitercutoff}{a numeric number indicating the log-likelihood cutoff applied to stop the EM algorithm}

\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{design:}{a matrix. Number of rows should be the same as the number of unique samples. Rownames are sample names. First column is the intercept (all 1), second column is the covariate realization valuels for each of the samples.}
}
\value{
a list containing parameters in the model 0 for genes.
}
\description{
This function is used to perform the model 0 (beta0 for intercept base only) fitting.
}
\examples{
data(mandata)
a = fitpt.m0(expr = mandata$expr, cellanno = mandata$cellanno, pseudotime = mandata$pseudotime, design = mandata$design, EMmaxiter=5, EMitercutoff=10)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
