% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDDGType.R
\name{getDDGType}
\alias{getDDGType}
\title{Obtain the DDG types for the DDG in sample covariate test.}
\usage{
getDDGType(testobj, cutoff = 0.05)
}
\arguments{
\item{testobj}{output object of the function tespt().}

\item{cutoff}{the FDR cutoff to select "statistically significant" genes.}
}
\value{
a vector of DDGType. The names of this vector are the DDG.
}
\description{
This function is used to obtain the DDG types for the DDG in sample covariate test.
}
\examples{
data(mantestobj)
a = getDDGType(mantestobj)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
