% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDDGType.R
\name{getXDEType}
\alias{getXDEType}
\title{Obtain the DDG types for the DDG in sample covariate test.}
\usage{
getXDEType(testobj, cutoff = 0.05)
}
\arguments{
\item{testobj}{output object of the function tespt().}

\item{cutoff}{the FDR cutoff to select "statistically significant" genes.}
}
\value{
a vector of XDE Types, including meanSig, trendSig, bothSig, other, and nonXDE. The names of this vector are the XDE genes.
}
\description{
This function is used to obtain the DDG types for the DDG in sample covariate test.
}
\examples{
data(mantestobj)
a = getXDEType(mantestobj)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
