% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOEnrich.R
\name{GOEnrich}
\alias{GOEnrich}
\title{Obtain enrich Gene Ontology (GO) terms}
\usage{
GOEnrich(
  testobj,
  fdr.cutoff = 0.05,
  k = 5,
  use.clusters = TRUE,
  type = "Variable",
  species = "human",
  sep = ":.*"
)
}
\arguments{
\item{testobj}{output object of lamian_test()}

\item{fdr.cutoff}{FDR cutoff to select "statistically significant" GO terms.}

\item{k}{number of clusters to be used for clustering, if the gene clustering results have not yet been included in testobj.}

\item{use.clusters}{If TRUE (default), use the clusters in testobj.}

\item{type}{the type of the differential test. One of c('Time', 'Variable'). Use for selecting the columns in statistics of FDR. Only useful when "cluster" in not in testobj.}

\item{species}{currently only work for "human". will include "mouse".}
}
\value{
a list. Each element is a data frame about the GO terms and their statistics.
}
\description{
This function is used to obtain enrich Gene Ontology (GO) terms (currently only for human genes)
}
\examples{
data(mantestobj)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
