% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_tree_structure.R
\name{infer_tree_structure}
\alias{infer_tree_structure}
\title{Infer the pseudotime tree strucutre of the input data}
\usage{
infer_tree_structure(
  pca,
  cellanno,
  expression,
  origin.marker = NA,
  origin.celltype = NA,
  number.cluster = NA,
  plotdir = NA,
  xlab = "PC1",
  ylab = "PC2",
  max.clunum = 50,
  kmeans.seed = 12345
)
}
\arguments{
\item{pca}{cell by principal component (pc) matrix. Principal components reduction of the cells.}

\item{cellanno}{2-column or 3-column dataframe/matrix, first column is cell name, second column is sample, third column (if exists) is cell type.}

\item{expression}{only useful when users want to use highly expressed marker genes to determine the starting point of pseudotime. It is a gene by cell expression matrix. The values should be library-size-normalized and log-transformed expression values. They can either be imputed or non-imputed.}

\item{origin.celltype}{only useful when users have specify the celltypes for the cells in cellanno[,3], and want to determin the starting point of pseudotime by this celltype.  It is a character of origin celltype. It should be one of the element in cellanno[,3] if it will be used. If NA (default), then not used.}

\item{number.cluster}{the number of clusters in cell clustering that will be used in trajectory inference. If NA (default),the number of clusters will be determined automatically by elbow's method.}

\item{plotdir}{plot directory for storing the figures of cluster.pdf and pseudotime.pdf. If NA (default), figures will not be generated.}

\item{xlab}{the x-axis labels for the figures.}

\item{ylab}{the y-axis labels for the figures.}

\item{max.clunum}{the maximum number of clusters in the elbew's method.}

\item{origin.marker:}{a character or a character vector specifying the highly expressed marker genes used to identify the origin cluster in pseudotime inference.If NA (default), then not used.}
}
\value{
a list
}
\description{
This function takes a low dimensional-reduction representation (for example, pca), the cell annotation, etc. as inputs, and then infer the pseudotime tree structures for further assessing the uncertainty of each of the pseudotime tree branches.
}
\examples{
data(man_tree_data)
res = infer_tree_structure(pca = man_tree_data[['pca']], cellanno = man_tree_data[['cellanno']], expression = man_tree_data[['expression']], origin.marker = c('CD34'), number.cluster = 5, xlab='Principal component 1', ylab = 'Principal component 2')
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
