% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamian_test_h5.R
\name{lamian_test_h5}
\alias{lamian_test_h5}
\title{Test function for all differential tests in lamian module 2 when the input gene expression is in hdf5 format}
\usage{
lamian_test_h5(
  expr,
  cellanno,
  pseudotime,
  design = NULL,
  testvar = 2,
  permuiter = 100,
  EMmaxiter = 100,
  EMitercutoff = 0.05,
  verbose.output = F,
  ncores = detectCores(),
  test.type = "Time",
  fit.resolution = 1000,
  return.all.data = TRUE,
  overall.only = F,
  test.method = "permutation",
  ncores.fit = 1,
  fix.all.zero = TRUE,
  cutoff = 1e-05
)
}
\arguments{
\item{expr}{hdf5 path dir (including file name).}

\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{testvar}{a numeric number indicating the column in the design matrix that needs to be tested while controlling for other columns (not intercept). Default is 2. testvar = 2 means the second column in the design needs to be tested.}

\item{permuiter}{an integer  indicating the number of permutations performed in the permutation test.}

\item{EMmaxiter}{an integer indicating the number of iterations in the EM algorithm.}

\item{EMitercutoff}{a numeric number indicating the log-likelihood cutoff applied to stop the EM algorithm}

\item{verbose.output}{logical. If TRUE, print intermediate information.}

\item{ncores}{number of cores for performing the permutation test.}

\item{test.type}{One of c('Time', 'Variable'). Case insensitive.}

\item{return.all.data}{logical. If TRUE (default), return all data including inputs.}

\item{overall.only}{logical. If TRUE (default), only test the overall fdr and skip the trend fdr and mean fdr. Default is FALSE.}

\item{test.method}{One of c('chisq', 'permutation). If 'permutation' (default), use the permutation test to identify genes. If 'chisq', use the chisq test to identify genes.}

\item{ncores.fit}{is the ncores for fitpt_h5() or fitfunc_h5()(essentially fitpt_h5()) only. It only works when test.method = 'chisq'.}

\item{fix.all.zero}{logical. If TRUE (defalt), fix the issue of all zeros in any samples.}

\item{cutoff}{a numeric number to set the cutoff for the standard deviation of gene expression in any one of the samples. Only useful when fix.all.zero == TRUE.}

\item{design:}{a matrix. Number of rows should be the same as the number of unique samples. Rownames are sample names. First column is the intercept (all 1), second column is the covariate realization valuels for each of the samples.}
}
\value{
a list of results
}
\description{
This function is designed for doing different tests when the input gene expression is in hdf5 format.
}
\examples{
\dontrun{
data(expdata)
res <- lamian_test_h5(expr='multi.h5', cellanno=expdata$cellanno, pseudotime=expdata$pseudotime, design=expdata$design, testvar=2, test.type = 'Variable', overall.only = F, test.method = 'chisq')
}
data(mandata)
a = lamian_test(expr = mandata$expr[seq(1,3),], cellanno = mandata$cellanno, pseudotime = mandata$pseudotime, design = mandata$design, test.method = 'chisq', permuiter = 10, EMmaxiter = 10, EMitercutoff = 10, verbose.output = FALSE)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
