% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykmeans.R
\name{mykmeans}
\alias{mykmeans}
\title{Perform k-means clustering}
\usage{
mykmeans(matrix, number.cluster = NA, maxclunum = 30, seed = 12345)
}
\arguments{
\item{matrix}{cell by principal component (pc) matrix for cell clustering.}

\item{number.cluster}{the number of clusters in cell clustering that will be used in trajectory inference. If NA (default),the number of clusters will be determined automatically by elbow's method.}

\item{maxclunum}{the maximum number of clusters in the elbew's method.}

\item{seed}{seed applied before kmeans() to address it reproducibility issue.}
}
\value{
k-means clustering result
}
\description{
Perform k-means clustering for the cells (or genes) on any cell (or gene) by low-dimensional reprensentation coordinates matrix.
}
\examples{
data(mandata)
a = mykmeans(matrix = mandata$expr, number.cluster = 2)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
