% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterDiff.R
\name{plotClusterDiff}
\alias{plotClusterDiff}
\title{Plot the cluster mean or of group differnce for a set of genes or plot the group difference for individual genes.}
\usage{
plotClusterDiff(
  testobj,
  gene = names(testobj$cluster),
  cluster = testobj[["cluster"]],
  each = FALSE,
  sep = "",
  reverse = FALSE,
  facet.grid = FALSE,
  free.scale = TRUE,
  axis.text.blank = FALSE
)
}
\arguments{
\item{testobj}{output object from lamian_test(). It is a list.}

\item{gene}{a vector of gene names.}

\item{cluster}{users can defined the clusters of the cells. It is a numeric vector whose names are cell names. By default it is retrieved from testobj object.}

\item{each}{logical. If TRUE, the plot for each gene will be generated.}

\item{sep}{a string in the gene names that needs to replaced with blank.}

\item{reverse}{logical. If FALSE (default), group difference is calculated by group 1 - group 0. If TRUE, group difference is calculated by group 0 - group 1.}

\item{facet.grid}{logical. If FALSE (default), seperate clusters by facet_wrap, otherwise by facet_grid.}

\item{free.scale}{logical. If TRUE (default), the y-axis is on free scale.}

\item{axis.text.blank}{logical. If TRUE, leave axis text as blank.}
}
\value{
a plot
}
\description{
This function is used for plotting the cluster mean of group difference or plotting the group difference for individual genes.
}
\examples{
data(mantestobj)
plotClusterDiff(testobj = mantestobj)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
