% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterMean.R
\name{plotClusterMean}
\alias{plotClusterMean}
\title{Plot the cluster mean of population-level fitting values.}
\usage{
plotClusterMean(testobj, cluster, type = "time", facet = FALSE)
}
\arguments{
\item{testobj}{output object from lamian.test(). It is a list.}

\item{cluster}{users can defined the clusters of the cells. It is a numeric vector whose names are cell names. By default it is retrieved from testobj object.}

\item{type}{one of c('Time', 'Variable').}

\item{facet}{logical. If TRUE, facet each cluster.}
}
\value{
a plot
}
\description{
This function is used for plotting the cluster mean of population-level fitting values.
}
\examples{
data(mantestobj)
plotClusterMean(mantestobj, type = 'variable')
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
