% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterMeanAndDiff.R
\name{plotClusterMeanAndDiff}
\alias{plotClusterMeanAndDiff}
\title{Plot the cluster mean of population-level fitting values and group difference.}
\usage{
plotClusterMeanAndDiff(
  testobj,
  cluster = testobj[["cluster"]],
  free.scale1 = TRUE,
  free.scale2 = FALSE
)
}
\arguments{
\item{testobj}{output object from lamian_test(). It is a list.}

\item{cluster}{users can defined the clusters of the cells. It is a numeric vector whose names are cell names. By default it is retrieved from testobj object.}

\item{free.scale1}{logical. If TRUE, free scale the y-axis of the cluster mean plot.}

\item{free.scale2}{logical. If TRUE, free scale the y-axis of the group difference plot.}
}
\value{
a plot
}
\description{
This function is used for plotting the cluster mean of population-level fitting values and group difference.
}
\examples{
data(mantestobj)
plotClusterMeanAndDiff(testobj = mantestobj)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
