% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCellAndPopulation.R
\name{plotGeneCellAndPopulation}
\alias{plotGeneCellAndPopulation}
\title{Plot genes' cell and population-level fitting.}
\usage{
plotGeneCellAndPopulation(
  testobj,
  gene = NA,
  type = "time",
  ylim = NA,
  xlim = NA,
  sep = NA,
  free.scale = TRUE,
  palette = "Dark2",
  ncol = NA,
  subSampleNumber = NA,
  line.size = 1,
  line.alpha = 1,
  point.size = 0.2,
  point.alpha = 0.2,
  axis.text.blank = FALSE,
  dot.quantile = 1
)
}
\arguments{
\item{testobj}{object returned from lamian_test().}

\item{gene}{a character vector of gene names. It can be of length 1 or > 1.}

\item{type}{One of c('Time', 'Variable').}

\item{ylim}{y-axis limits to be passed to ggplot.}

\item{xlim}{x-axis limits to be passed to ggplot.}

\item{sep}{a string in the gene names that needs to replaced with blank.}

\item{free.scale}{logical. If TRUE, the y-axis is on free scale.}

\item{palette}{a RColorBrewer palette name.}

\item{ncol}{number of colums for organizing all genes' plot.}

\item{subSampleNumber}{a numeric number. If specified, use it as the number of cells subsampled from all cells that will be shown in the plot}

\item{line.size}{the size of the curves.}

\item{point.size}{the size value of points.}

\item{point.alpha}{the alpha value of points.}

\item{axis.text.blank}{logical. If TRUE, leave axis text as blank.}
}
\value{
a plot
}
\description{
This function is used to plot genes' using their cellular level expression and the population-level fitting values.
}
\examples{
data(mantestobj)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
