% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneSampleAndPopulation.R
\name{plotGeneSampleAndPopulation}
\alias{plotGeneSampleAndPopulation}
\title{Plot genes' sample-level and population-level fitting plot for sample covariate test.}
\usage{
plotGeneSampleAndPopulation(
  testobj,
  gene,
  variable = NULL,
  variable.text = NULL,
  free.scale = TRUE,
  facet.sample = FALSE,
  plot.point = FALSE,
  line.alpha = 1,
  line.size = 1,
  point.alpha = 1,
  point.size = 0.5,
  continuous = TRUE,
  sep = NA,
  palette = "Dark2",
  ncol = NULL,
  axis.text.blank = F
)
}
\arguments{
\item{gene}{a vector of genes that need to do the prediction.}

\item{variable}{character, the variable (covariate) to color the samples, should be null or one of the column names of design matrix. Default is NULL, meaning each sample is colored differently. Otherwise, samples are colored by the variable (covariate) values.}

\item{variable.text}{a character vector. The text for the legend of the plot, corresponding to each variable values.}

\item{free.scale}{logical. If TRUE, the y-axis is on free scale.}

\item{facet.sample}{logical. If TRUE, each sample is plotted separately.}

\item{plot.point}{logical. If TRUE, cells will be plotted as ponts.}

\item{line.alpha}{alpha value for lines.}

\item{line.size}{size value for lines.}

\item{point.alpha}{alpha value for points.}

\item{point.size}{size value for points.}

\item{sep}{a string in the gene names that needs to replaced with blank.}

\item{palette}{a RColorBrewer palette name}

\item{ncol}{number of columns}

\item{axis.text.blank}{logical. If TRUE, leave axis text as blank.}

\item{testObj}{the output object from lamian.test().}

\item{continuous:}{if TRUE, samples are colored using viridis continuous colors. If FALSE, RColorBrewer "Dark2" discrete palette.}

\item{expression}{a character ('demean',or 'original') to define the expression values shown on the plots. if "demean"(default), show demeaned expression. if 'original", show original gene expression.}
}
\value{
a plot
}
\description{
This function is used to plot genes' sample-level and population-level fitting plot for sample covariate test.
}
\examples{
data(mantestobj)
plotGeneSampleAndPopulation(testobj = mantestobj, gene = rownames(mantestobj$populationFit[[1]])[1], variable = 'gender')
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
