% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOEnrich.R
\name{plotGOEnrich}
\alias{plotGOEnrich}
\title{Plot enriched GO terms for each of the clusters.}
\usage{
plotGOEnrich(goRes, n = 5, sortByFDR = TRUE, fdr.cutoff = 0.05, fc.cutoff = 2)
}
\arguments{
\item{goRes}{output from GOEnrich. A list of GO enrichment result. Length same as number of clusters.}

\item{n}{number of top GO terms shown for each cluster}

\item{sortByFDR}{if TRUE (default), sort GO terms by FDR first, and then negative FC. If FALSE,  soft by negative FC and then FDR.}

\item{fdr.cutoff}{FDR cutoff for choosing "statistically significant" GO terms. Will be combined with those choosing by fc.cutoff.}

\item{fc.cutff}{fold change cutoff for choosing "statistically significant" GO terms. Will be combined with those choosing by fdr.cutoff.}
}
\value{
a plot
}
\description{
This function is used to plot enriched GO terms for each of the clusters.
}
\examples{
data(mantestobj)
plotGOEnrich(goRes = mantestobj$goRes, n = 2, fdr.cutoff = 1, fc.cutoff = 2)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
