% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmclust.R
\name{plotmclust}
\alias{plotmclust}
\title{Plot the pseudotime tree structure. Overwrite TSCAN::plotmclust().}
\usage{
plotmclust(
  mclustobj,
  x = 1,
  y = 2,
  MSTorder = NULL,
  show_tree = TRUE,
  show_full_tree = TRUE,
  show_cell_names = FALSE,
  cell_name_size = 3,
  cell_point_size = 3,
  markerexpr = NULL,
  showcluster = TRUE,
  x.lab = "PC1",
  y.lab = "PC2",
  subset.cell = NULL
)
}
\arguments{
\item{x}{indicates which column in the low-dimrension representation to be used as x-axis coordinates.}

\item{y}{indicates which column in the low-dimrension representation to be used as y-axis coordinates.}

\item{MSTorder}{The arbitrary order of cluster to be shown on the plot.}

\item{show_tree}{Whether to show the links between cells connected in the minimum spanning tree.}

\item{show_full_tree}{Whether to show the full tree or not. Only useful when show_tree=TRUE. Overrides MSTorder.}

\item{show_cell_names}{Whether to draw the name of each cell in the plot.}

\item{cell_name_size}{The size of cell name labels if show_cell_names is TRUE.}

\item{cell_point_size}{The size of cell point.}

\item{markerexpr}{The gene expression used to define the size of nodes.}

\item{showcluster}{logical. indicates whether to show clusters on the plot.}

\item{x.lab}{x-axis label}

\item{y.lab}{y-axis label}

\item{subset.cell}{a character vector of the names of the selected cells to plot, when NULL, plot all cells.}
}
\value{
A ggplot2 plot
}
\description{
This function takes the output from infer_tree_structure() to generate a plot. For example, if cells are using pc as dimension reduction, then the plot will be y-axis as PC2, a-axis as PC1 (or otherwise specified by inputs x and y), and dots are cells colored by clusters. A pseudotime tree structure is shown on top of the cells.
}
\examples{
data(man_tree_res)
plotmclust(man_tree_res)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
