% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fitting.R
\name{predict_fitting}
\alias{predict_fitting}
\title{Predict population fitting.}
\usage{
predict_fitting(testObj, gene = NULL, test.type = "time")
}
\arguments{
\item{testObj}{the output object from lamian_test().}

\item{gene}{a vector of genes that need to do the prediction.}

\item{test.type}{One of c('Time', 'Variable') and case-insentivie.}
}
\value{
a gene by cell (or pseudotime) expression matrix
}
\description{
This function will predict the population fitting for genes using the output from lamian_test().
}
\author{
Wenpin Hou <wp.hou3@gmail.com>
}
