% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalematrix.R
\name{scalematrix}
\alias{scalematrix}
\title{Standardize the values by rows}
\usage{
scalematrix(data)
}
\arguments{
\item{data}{a matrix.}
}
\value{
a gene by cell (or pseudotime) expression matrix
}
\description{
This function standardize the values for a matrix by its rows
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
