% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensFdr.R
\name{SensFdr}
\alias{SensFdr}
\title{Calculate Sensitivity, real FDR, and reported FDR}
\usage{
SensFdr(TruePositive, Statistics)
}
\arguments{
\item{TruePositive}{a character vector of true positive genes.}

\item{Statistics:}{a dataframe or matrix, should contain a column of fdr , for example c('FStat','P.Value','adj.P.Val'). Row names are gene names.}
}
\value{
a datafrome. Columns are c('Sensitivity','Real_FDR','Reported_FDR').
}
\description{
This function is used to calculate Sensitivity, real FDR, and reported FDR.
}
\examples{
stat = data.frame(fdr = abs(rnorm(10)), fc = rnorm(10))
rownames(stat) = letters[seq(1,10)]
a = SensFdr(TruePositive = letters[seq(1,5)], Statistics = stat)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
