% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gene.R
\name{cluster_gene}
\alias{cluster_gene}
\title{Support function for external clusterGene(). Cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).}
\usage{
cluster_gene(
  testobj,
  gene,
  k,
  k.auto = FALSE,
  type = "Time",
  method = "kmeans",
  scale.difference = F,
  seed = 12345
)
}
\arguments{
\item{testobj}{object returned from lamian.test().}

\item{gene}{a character vector of gene names. It can be of length 1 or > 1.}

\item{k}{a numeric number. The number of clusters. Only useful when k.auto = FALSE.}

\item{k.auto}{logical. If FALSE (default), users need to specify k as the number of clusters. If TRUE, k will be automatically determined by elbow's method.}

\item{type}{One of c('Time', 'Variable').}

\item{method}{The clustering method. "kmeans" (default) for k-means clustering,  "hierarchical" for hierarchical clustering, "louvain" for Louvain clustering, and "GMM" for Model-based clustering.}

\item{scale.difference}{logical. If FALSE, then do not standarize the group difference, but only scale by the maximum value of the group difference absolute values. If TRUE, then standardize the group difference before doing the clustering.}
}
\value{
a plot
}
\description{
This function is used to support the external function clusterGene(). cluster genes based on their temporal patterns  of gene expression (constant time test) or the temporal patterns of gene expression group difference (sample covariate test).
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
