% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_uncertainty.R
\name{evaluate_uncertainty}
\alias{evaluate_uncertainty}
\title{Evaluate the pseudotime tree uncertainty}
\usage{
evaluate_uncertainty(
  inferobj,
  n.permute,
  subset.cell = NULL,
  design = NULL,
  return.ctcomp = FALSE
)
}
\arguments{
\item{inferobj}{the output object from function infer_tree_structure().}

\item{subset.cell}{a character vector of the names of the selected cells where boostrap will happen on. If NULL, then boostrap from all the cells.}

\item{design}{a design matrix for performing branch proportion test. Each row is a sample. First column is intercept (all 1). Second column is covariate.}

\item{return.ctcomp}{if TRUE, return all branch proportion values of permutations. The default is FALSE.}

\item{n.permute:}{a numeric number of permutation in the permutation test.}
}
\value{
a list of results
}
\description{
This function is designed to evaluate the pseudotime tree uncertainty, as one of the main functions in lamain module 1.
}
\examples{
data(man_tree_res)
a <- evaluate_uncertainty(inferobj = man_tree_res, n.permute = 3)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
