% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovariateGroupDiff.R
\name{getCovariateGroupDiff}
\alias{getCovariateGroupDiff}
\title{Obtain the covariate group difference of the genes in sample covariate test.}
\usage{
getCovariateGroupDiff(testobj, gene, reverse = FALSE)
}
\arguments{
\item{testobj}{output object of the function tespt().}

\item{gene}{a character vector of genes.}

\item{reverse}{logitcal. FALSE (default) when group 1 - group 0.  TRUE when group 0 - group 1.}

\item{output}{a gene by pseudotime matrix. Entries are group difference w.r.t the variable, i.e., the unit-covariate incremental difference.}
}
\value{
a matrix of gene by cell (ordered by pseudotime). Values are covariate group difference.
}
\description{
This function is used to obtain the covariate group difference of the genes in sample covariate test.
}
\examples{
data(mantestobj)
a = getCovariateGroupDiff(testobj = mantestobj, gene = rownames(expr)[1:2])
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
