% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulationFit.R
\name{getPopulationFit}
\alias{getPopulationFit}
\title{Obtain the population-level fitting of the genes.}
\usage{
getPopulationFit(testobj, gene = NULL, type = "time", num.timepoint = 1000)
}
\arguments{
\item{testobj}{output object of the function tespt().}

\item{gene}{a vector of gene names}

\item{type}{one of c('Time', 'Variable). Case insensitive.}

\item{num.timepoint}{the number of time points used to fit the curve. Default is the minimum of 1e3 and max(pseudotime). This argument will reduce the running time of both this function and the downstream plot function if there are too many cells.}
}
\value{
a matrix, population fitting of the genes (rows) by cells (ordered by pseudotime)
}
\description{
This function is used to obtain the population-level fitting pattern along pseudotime of the genes, using the estimated parameters from the lamian model.
}
\examples{
data(mantestobj)
a <- getPopulationFit(testobj = mantestobj, gene = rownames(mantestobj$populationFit[[1]])[seq(1,3)], type = 'variable')
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
