% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{Plot gene(s) by showing both of the original cellcular gene expression and the sample-level, population-level fitting values.}
\usage{
plotGene(
  testobj,
  gene,
  variable = NULL,
  variable.text = NULL,
  free.scale = TRUE,
  facet.sample = FALSE,
  plot.point = FALSE,
  line.alpha = 1,
  line.size = 1,
  point.alpha = 1,
  point.size = 0.5,
  continuous = TRUE,
  sep = NA,
  palette = "Dark2",
  ncol = NULL,
  axis.text.blank = FALSE,
  cellProp = FALSE,
  x.lab = "Pseudotime",
  y.lab = "Expression"
)
}
\arguments{
\item{testobj}{object returned from lamian_test().}

\item{gene}{a character vector of gene names. It can be of length 1 or > 1.}

\item{variable}{character, the variable (covariate) to color the samples, should be null or one of the column names of design matrix. Default is NULL, meaning each sample is colored differently. Otherwise, samples are colored by the variable (covariate) values.}

\item{variable.text}{a character vector. The text for the legend of the plot, corresponding to each variable values.}

\item{free.scale}{logical. If TRUE, the y-axis is on free scale.}

\item{facet.sample}{logical. If TRUE (default), facet_wrap the samples.}

\item{plot.point}{point size}

\item{line.alpha}{alpha value of the curves}

\item{line.size}{the size of the curves.}

\item{point.alpha}{the alpha value of points.}

\item{point.size}{the size value of points.}

\item{continuous}{if TRUE, samples are colored using viridis continuous colors. If FALSE, RColorBrewer "Dark2" discrete palette.}

\item{sep}{a string in the gene names that needs to replaced with blank.}

\item{palette}{a RColorBrewer palette name.}

\item{ncol}{number of colums for organizing all genes' plot.}

\item{axis.text.blank}{logical. If TRUE, leave axis text as blank.}

\item{cellProp}{logical. If FALSE (default), plot gene expression. If TRUE, it is cell proportion.}

\item{x.lab}{a string to indicates x-axis label}

\item{y.lab}{a string to indicates y-axis label}

\item{type}{One of c('Time', 'Variable').}

\item{ylim}{y-axis limits to be passed to ggplot.}

\item{xlim}{x-axis limits to be passed to ggplot.}
}
\value{
a plot
}
\description{
This function is used to plot gene(s) by showing both of the original cellcular gene expression and the sample-level, population-level fitting values.
}
\examples{
data(mantestobj)
plotGene(testobj = mantestobj, gene = rownames(mantestobj$populationFit[[1]])[1], variable = 'gender')
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
