% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTDEHm.R
\name{plotTDEHm}
\alias{plotTDEHm}
\title{Plot the fitting heatmaps for TDE test.}
\usage{
plotTDEHm(
  testobj,
  showRowName = FALSE,
  cellWidthTotal = 250,
  cellHeightTotal = 400,
  showCluster = FALSE,
  colann = NULL,
  rowann = NULL,
  annotation_colors = NULL,
  type = "time",
  subsampleCell = TRUE,
  numSubsampleCell = 1000
)
}
\arguments{
\item{testobj}{output object from lamian_test(). It is a list.}

\item{showRowName}{logical. If FALSE (default), row names (i.e. gene names) of the heatmaps will not be shown in the plot.}

\item{cellWidthTotal}{a numeric number. Total width of each heatmap cell.}

\item{cellHeightTotal}{when showRowName = TRUE, cellHeightTotal is suggested to be ten times the number of genes (rows).}

\item{showCluster}{(no implemented yet). if TRUE, "cluster" should be a slot in testobj, and it will be label in the heatmap. If FALSE, no need to pass in "cluster".}

\item{colann}{a data frame. Each column represent the annotation feature for the cells. row names should be the same as the name of the cells in the data.}

\item{rowann}{a data frame. Each column represent the annotation feature for the genes. row names should be the same as the name of the genes in the data.}

\item{annotation_colors}{a list. Will be passed onto the annotation_colors input in pheatmap().}

\item{type}{One of c('Time', 'Variable')}

\item{subsampleCell}{logical. If TRUE, will subsample cells.}

\item{numSubsampleCell}{a numeric number indicating the number of cells users want to subsampled. Only useful when subsampleCell == TRUE.}
}
\value{
a plot
}
\description{
This function is used for plotting the fitting heatmaps.
}
\examples{
data(mantestobj)
plotTDEHm(testobj = mantestobj, type = 'variable', numSubsampleCell = 5)
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
