% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveh5.R
\name{saveh5}
\alias{saveh5}
\title{This function generates an hdf5 file as the gene expression input to lamian_test_h5().}
\usage{
saveh5(
  expr,
  pseudotime,
  cellanno,
  path,
  fix.all.zero = TRUE,
  cutoff = 0.001,
  sd.adjust = 0.001
)
}
\arguments{
\item{expr}{gene by cell expression matrix. Values are library-size-normalized log-transformed gene expression matrix. They can be either imputed or non-imputed. Zero-expression genes should have been filtered out.}

\item{pseudotime}{a numeric vector of pseudotime, and the names of this vector are the corresponding cell names.}

\item{cellanno}{a dataframe where the first column are cell names and second column are sample names.}

\item{path}{a path (string, with the file name suffix .h5) to write three files. An example is 'data/multi.h5'.}

\item{fix.all.zero}{logical. If TRUE (default), fix the case when expression values  within a sample is too small (smaller than  the cutoff).}

\item{cutoff}{a numeric value. The cutoff used to determine whether to add some very small white noise to genes when fix.all.zero is TRUE.}

\item{sd.adjust}{the variance of the noise to be added if fix.all.zero is TRUE.}
}
\value{
No return but write an hdf5 file into a path.
}
\description{
This function generates an hdf5 file as the gene expression input to lamian_test_h5().
}
\examples{
data(expdata)
saveh5(expr = expdata$expr, pseudotime = expdata$pseudotime, cellanno = expdata$cellanno, path = 'data/multi.h5')
}
\author{
Wenpin Hou <whou10@jhu.edu>
}
